tlparser/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Parser for Twich streamers and games and Lamoda products"
authors = ["Twelve <kana.suzucki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
flake8 = "6.1.0"
black = "23.9.1"
isort = "5.12.0"
mypy = "1.5.1"
Flake8-pyproject = "1.2.3"
fastapi = "0.104.1"
pydantic = "2.5.2"
pydantic-settings = "2.1.0"
uvicorn = "0.24.0"
dependency-injector = "4.41.0"
pymongo = "4.6.1"
mongoengine = "0.27.0"
redis = "5.0.1"
types-redis = "4.6.0.11"
fastapi-cache2 = "0.2.1"
kafka-python = "2.0.2"
elasticsearch="7.10.1"
elasticsearch-dsl = "7.4.0"
aiohttp =  "3.9.1"
py-automapper = "1.2.3"
strawberry-graphql = {extras = ["fastapi"], version = "=0.219.2"}


[tool.poetry.group.prd]
optional = true

[tool.poetry.group.prd.dependencies]


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]


[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "7.2.0"
pytest-cov = "4.0.0"
pytest-mock = "3.10.0"
httpx = "0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 100
lines_before_imports = 2
lines_after_imports = 2
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
treat_all_comments_as_code = true
lines_between_sections = 0
split_on_trailing_comma = true

[tool.flake8]
max-line-length = 100
exclude = ['migrations', 'data', '.venv']
ignore = ['E203', 'E266', 'E704', 'W503', 'B905', 'B907']
max-complexity = 18
select = ['B', 'C', 'E', 'F', 'W', 'T4', 'B9']

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
strict_optional = true
exclude = '''(?x)(
    data/$
    | migrations/$
    | tests/$
)'''
plugins = [
    "pydantic.mypy"
]

[tool.black]
line-length = 100
skip-string-normalization = true
preview = true
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
)/
'''

[tool.pytest.ini_options]
minversion = "7.2.0"
testpaths = "tests"
norecursedirs = ".venv .mypy_cache .pytest_cache data docs"
addopts = "-l -v -rsxX -p no:warnings --tb=short --strict-markers"
python_files = "test_* *_test tests_* *_tests unit* *unit func* *func"
python_classes = "*Test Test*"
python_functions = "test_*  *_test"

[tool.coverage.run]
branch = true
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*apps.py",
    "*settings*.py",
    "*urls.py",
    "*wsgi.py",
    "*asgi.py",
    "manage.py"
]

[tool.coverage.report]
show_missing = true

[tool.coverage.html]
directory = ".coverage_report"